Implement conforming Delaunay triangulation by inserting more points when adding constraints
Ensure that triangulation conforms to constraints (fixed edges)
For each fixed edge that is not present in the triangulation its
midpoint is recursively added until the original edge is represented by a
sequence of its pieces.
New vertices are inserted.

One important modification compared to Triangle: allow to flip already constraint edges and re-add them later.
This produces a truly Delaunay triangulation as a result.

Overlapping boundaries will work with conform-to-edge: test file added.