#204 Fix inserting vertex directly on fixed edge
When interleaving multiple calls to `insertVertices` and `insertEdges` it is possible that the vertex splits fixed edge.
The code did not handle the case properly: triangles were split but fixed edges were not updated.

Fix: when inserting vertex on the edge check if edge is fixed and do all the necessary bookkeeping.

Regression test added.